src/ui/server/controllers/insurance/export-contract/map-and-save/private-market/index.ts
import hasFormData from '../../../../../helpers/has-form-data';
import save from '../../save-data/private-market';
import { Application, RequestBody, ValidationErrors } from '../../../../../../types';

/**
 * mapAndSave
 * Map and save any valid private market fields
 * @param {Express.Request.body} Express request body
 * @param {Application}
 * @param {Object} Validation errors
 * @returns {Boolean}
 */
const privateMarket = async (formBody: RequestBody, application: Application, validationErrors?: ValidationErrors) => {
  try {
    if (hasFormData(formBody)) {
      let saveResponse;

      if (validationErrors) {
        saveResponse = await save.privateMarket(application, formBody, validationErrors.errorList);
      } else {
        saveResponse = await save.privateMarket(application, formBody);
      }

      if (!saveResponse) {
        return false;
      }

      return true;
    }

    return true;
  } catch (err) {
    console.error('Error mapping and saving application %O', err);

    return false;
  }
};

export default {
  privateMarket,
};
